backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
langchain==0.1.1
langchain-community==0.0.13
langchain-openai==0.0.5
langchain-chroma>=0.0.5
chromadb==0.4.22
langchain-qdrant>=0.2.0
chroma-hnswlib>=0.7.3
python-dotenv==1.0.0
bcrypt==4.1.2
sqlalchemy==2.0.25
alembic>=1.12.1
mysql-connector-python==8.2.0
minio==7.2.0
python-docx>=0.8.11
pypdf>=3.0.0
unstructured==0.11.6
markdown>=3.0.0
unstructured[md]>=0.10.0
openai>=1.30.0
email-validator
python-magic==0.4.27
pdf2image==1.16.3
pytesseract==0.3.10
tiktoken==0.5.2